fix: agressive links, force override
I don't know how I feel about this
but given the amount of errors i've gotten when testing the laptop?
Some files already exist like

but what if you didn't want to overwrite it exactly?
dotbot doesn't have a very good way of dealing with this out of the box,
two options I have thought of are either,

- Backup plugin for dotbot
- Merge tools, maybe utilize git somehow?