feat: pwsh compatible bootstrap script
I modified the installation script to run as either bash or pwsh.

While this allows for more shells to be utilized, it is not without
it's issues.

Ideally, I'd like to be able to make my dotfiles OS agnostic,
but with the current level of cross shell safety dotbot offers,
this is more of a hack than anything.
Symlinks are created on the windows host but the cleanup throws an
invalid link error.