feat: install config for bpytop via pip
Basic meta/config file describing how to install bpytop system monitor
using python3 pip package. This is installed as a user package so the
path ~/.local/bin needs to be set in your $PATH.

The reason for using pip is because bpytop is not in included in the
ubuntu package repository, even though it is available in the Debian
repository.