refactor: move python_history to xdg_state_home
The preferred location is next to bash/zsh_history files.
pythonrc will attempt to set the location based off environment
variables, and will fallback to $HOME if the variable is not set.

These used to be either $HOME as dotfiles or in XDG_CACHE_HOME,
Though with the new XDG_STATE_HOME specification, I like the idea of
this being their new home.