refactor: zsh config deploy remove glob excludes
Allow explicitly linking important files. [zshenv, zshrc]
These link either outside ZDOTDIR or require a period.

Tab completing for zshrc can be confusing when working through
`~/.config/` instead of from the dotfiles repository.
And by having two symlinks when ignoring the excludes section:
We have a `.zshrc` that is properly loaded, and extra symlinks to both
files for ease of access to the user.