fix: overlooked bashrc.d glob pattern. prefer *.bash
This should have been changed in an earlier release but was missed.
treating bash shell scripts as `.conf` files is a bit weird and the
distinction is better like this if at all.

Might consider `.sh` in the future depending on how incompatible some
shell scripts are, I would like as much cross compatibility as possible.