refactor: minimize bashrc, debian common loaded from dynamic config
Most of these settings I'm not even sure what they do and at least one
states that it does not need to be enabled. Likely enabled already from
one of the `/etc/` scripts.

Ontop of that, a good amount of configuration has been moved to
`~/.commonrc` to load POSIX complaint cross shell settings.