refactor: add alias check for batcat utility
Semi-revert: 15aa4ec

Reintroduces command with proper check for `batcat` binary, instead of
`bat` shim.

When installing the utility 'bat: cat with wings', You should be able to
call the binary via `bat`. However, on ubuntu at the very least: The
command is `batcat`. In order to have consistency, we can check for
batcat and assign an alias for bat if it can be located.

Removes `cat` alias for less output, as it causes issue with the
standard needs for scripts that utilize it for concatenation.