fix: poor submodule handling w/ install script
The install script as well as the base config handles pulling submodule.
The `recursive` flag is required for dotbot libraries but results in all
submodules being pulled even if the user doesn't want to install
something like neovim and deploy the configuration repo.

We simply need to ensure a minimum environment for initializing any
packages and user config files.