refactor: retire `git br` custom style, prefer defaults
This alias acts as expected with your favorite flags, such as `-D`.
While the custom style displayed the recent commit message and relative
date that the commit was issued.

I like that functionality but when inspecting local branches it causes
some noise due to remote `origin/name` commits no longer are displayed
in red. Less useful for 'at-a-glance' work.