refactor: remove dynamic alias source from env script
Previously known as `~/.commonrc`, userenv defines some logic for
obtaining the current shell in order to source `~/.{shell}_aliases`.

This is undesirably complicated and would be better to let `~/.profile`
using `{shell}_VERSION` environment variable or the appropriate
`.{shell}rc` file to check that the alias file exists and source it.

tl;dr: Remove overly complicated and slow logic.