fix: disable shellcheck dynamic source in rc files
Relevant `.{}rc` files such as userenv, bash, zsh complain when sourcing
files dynamically or referencing a non-existent file to be sourced.
In these files, the content that is sourced has conditional checks to
ensure it's only sourced if the shell can locate them, thus the linting
error is to be expected but simply a distraction.

Adding the appropriate meta comments for shellcheck codes will disable
these across the whole file:

```bash
```