refactor: relocate user shell config to profile.d
In an attempt to further minimize and simplify deployment for both bash and zsh
settings, profile.d adopts the same approach used by the `/etc/profile.d/`
directory with a minor change. That being, the directory is sourced by both
`.bashrc` and `.zshrc` instead of `~/.profile`.

This is because these settings need to be reliably applied without a login shell
and allow the user to use the `alias=reload` in order to refresh their shell.

Configuration deployment for shellscript has been modified by merging and rename
some settings into `meta/config/profile.yaml`.