feat: add config/install for latest version of fzf
The apt version is way to old and causes some problems with the key
binding `Ctrl+r` to display completion history. This previously resulted
in an error due to an unrecognized command argument `--scheme=history`.

This adds an installation script which will clone the fzf repository to
the `$XDG_DATA_HOME/fzf` directory and run the provided install script.

Config has been packaged in its own directory and will be sourced by the
active shell via `profile.d/fzf.sh`.