feat(zsh): allow `ctrl-v` to edit the current command EDITOR
When a command gets out of hand, it's nice to be able to jump into your
editor and make more medications for newlines and indentation.

This adds a binding to quickly copy the content to a temporary file that
is opened in the currently configured EDITOR. Upon saving and exiting,
the command will be copied back into your active terminal session.