refactor(omp): change skellum right-side segment order
With this, segments are displayed in a mostly alphabetical order, with
the exception of a few select segments that are displayed in more
stylistic locations.

The top of the list are primary for utility segments for displaying
metadata for build/ci tools like Azure, AWS, Docker or any relevant
package/framework for the project within the current directory.

After which, the next section is for programming language. which display
information like semantic version numbers and/or virtual environments.