fix(omp): replace float to int on oh-my-posh v19.16.1
When Updating oh-my-posh through winget on my Windows 10 I received a config error. It seems that it is complaining about the typing for the config version number should be an integer instead of a semver style float

This seems like a weird change and I'm unsure if it holds up across builds for diffrent operating systems, so I will need to check before commiting to main.