fix(zsh): completion matcher causing mangling on invalid matches
Likely due to a bad regex string, any time the <tab> key is pressed when
none of the files match the current input, the preceding path will be
mangled with duplicate characters. This is extremely annoying, as it
requires the user to start over or jump backward and fix the start of
the path to continue.

In order to fix this, I simply replaced the zstyle with the exact one
found in the ohmyzsh/ohmyzsh repository. Specifically the hyphen
insensitive pattern.