fix(zsh): set explicit hl color flag for fzf-tab
Due to a hard-coded flag in the plugin when fzf is called, the
`--color=hl` value is set to some shade of white based on a condition.
This prevents the `hl` color value defined in `FZF_DEFAULT_OPTS` from
being utilized and requires the value to be set using `zstyle` flags for
the plugin.