perf(zsh): improve compinit loading conditions
Instead of allowing compinit to run on every new shell invocation, we
can compare the modified time and regenerate completion cache every 24
hours or exactly once per/day.

This however does have some drawbacks. Such as, when installing new
software that provides completions for zsh, completions are not easily
reloaded when restarting your shell. So, the user will likely need to
manually call compinit.

In future, it could be possible to check for changes in the fpath to
allow completions to be regenerated as needed. This will require more
research.