feat(zsh): exclude system users from completion
For common tasks, listing all of the system users just adds noise to the
completion results. By adding a number of ignore patterns, we can
simplify the user list to only show usernames we are interested in.

There my be cases where we want to modify one or more of these users but
I remain uncertain as to how common this is in practice. So for now,
we can keep this since the exact snippet is also present within the
oh-my-zsh base configuration.